backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.6.0
beanie==1.26.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.9
python-slugify==8.0.4
dnspython==2.6.1
